Adding more validations for ExerciseSet entity, and changing Float to BigDecimal, so the DB would recognize a decimal value after adding a @Digits annotation

+++ project/backend/src/main/java/com/exercise/project/dtos/ExerciseSetDTO.java
@@ -2,12 +2,14 @@ package com.exercise.project.dtos;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record ExerciseSetDTO(
         Long id,
         Long sets,
         Long reps,
-        Float weight,
+        BigDecimal weight,
         Long rest) {
 
 }
